for #2084, refactor package of ddl filler
