+++ sharding-core/sharding-core-parse/sharding-core-parse-common/src/main/java/org/apache/shardingsphere/core/parse/antlr/filler/common/ddl/column/RenameColumnDefinitionFiller.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.alter;
+package org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.column;
 
 import com.google.common.base.Optional;
 import lombok.Setter;
@@ -23,39 +23,28 @@ import org.apache.shardingsphere.core.metadata.table.ShardingTableMetaData;
 import org.apache.shardingsphere.core.parse.antlr.filler.api.SQLSegmentFiller;
 import org.apache.shardingsphere.core.parse.antlr.filler.api.ShardingTableMetaDataAwareFiller;
 import org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.column.ColumnDefinitionSegment;
-import org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.column.alter.ModifyColumnDefinitionSegment;
+import org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.column.alter.RenameColumnSegment;
 import org.apache.shardingsphere.core.parse.antlr.sql.statement.SQLStatement;
 import org.apache.shardingsphere.core.parse.antlr.sql.statement.ddl.AlterTableStatement;
 
 /**
- * Modify column definition filler.
+ * Rename column definition filler.
  *
  * @author duhongjun
  */
 @Setter
-public final class ModifyColumnDefinitionFiller implements SQLSegmentFiller<ModifyColumnDefinitionSegment>, ShardingTableMetaDataAwareFiller {
+public final class RenameColumnDefinitionFiller implements SQLSegmentFiller<RenameColumnSegment>, ShardingTableMetaDataAwareFiller {
     
     private ShardingTableMetaData shardingTableMetaData;
     
     @Override
-    public void fill(final ModifyColumnDefinitionSegment sqlSegment, final SQLStatement sqlStatement) {
+    public void fill(final RenameColumnSegment sqlSegment, final SQLStatement sqlStatement) {
         AlterTableStatement alterTableStatement = (AlterTableStatement) sqlStatement;
-        Optional<String> oldColumnName = sqlSegment.getOldColumnName();
-        if (oldColumnName.isPresent()) {
-            Optional<ColumnDefinitionSegment> oldColumnDefinition = alterTableStatement.findColumnDefinition(oldColumnName.get(), shardingTableMetaData);
-            if (!oldColumnDefinition.isPresent()) {
-                return;
-            }
-            oldColumnDefinition.get().setColumnName(sqlSegment.getColumnDefinition().getColumnName());
-            if (null != sqlSegment.getColumnDefinition().getDataType()) {
-                oldColumnDefinition.get().setDataType(sqlSegment.getColumnDefinition().getDataType());
-            }
-            alterTableStatement.getModifiedColumnDefinitions().put(oldColumnName.get(), oldColumnDefinition.get());
-        } else {
-            alterTableStatement.getModifiedColumnDefinitions().put(sqlSegment.getColumnDefinition().getColumnName(), sqlSegment.getColumnDefinition());
+        Optional<ColumnDefinitionSegment> oldColumnDefinition = alterTableStatement.findColumnDefinition(sqlSegment.getOldColumnName(), shardingTableMetaData);
+        if (!oldColumnDefinition.isPresent()) {
+            return;
         }
-        if (sqlSegment.getColumnPosition().isPresent()) {
-            alterTableStatement.getChangedPositionColumns().add(sqlSegment.getColumnPosition().get());
-        }
+        oldColumnDefinition.get().setColumnName(sqlSegment.getColumnName());
+        alterTableStatement.getModifiedColumnDefinitions().put(sqlSegment.getOldColumnName(), oldColumnDefinition.get());
     }
 }
