+++ sharding-core/sharding-core-parse/sharding-core-parse-common/src/main/resources/META-INF/parsing-rule-definition/common/filler-rule-definition.xml
@@ -17,15 +17,15 @@
   -->
 
 <filler-rule-definition>
-    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.column.ColumnDefinitionSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.ColumnDefinitionFiller" />
-    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.index.IndexSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.IndexFiller" />
-    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.constraint.ConstraintDefinitionSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.ConstraintDefinitionFiller" />
-    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.table.RenameTableSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.alter.RenameTableFiller" />
-    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.column.alter.AddColumnDefinitionSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.alter.AddColumnDefinitionFiller" />
-    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.column.alter.ModifyColumnDefinitionSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.alter.ModifyColumnDefinitionFiller" />
-    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.column.alter.RenameColumnSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.alter.RenameColumnDefinitionFiller" />
-    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.column.alter.DropColumnDefinitionSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.alter.DropColumnDefinitionFiller" />
-    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.constraint.DropPrimaryKeySegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.alter.DropPrimaryKeyFiller" />
+    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.column.ColumnDefinitionSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.column.ColumnDefinitionFiller" />
+    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.index.IndexSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.index.IndexFiller" />
+    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.constraint.ConstraintDefinitionSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.constraint.ConstraintDefinitionFiller" />
+    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.table.RenameTableSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.table.RenameTableFiller" />
+    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.column.alter.AddColumnDefinitionSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.column.AddColumnDefinitionFiller" />
+    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.column.alter.ModifyColumnDefinitionSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.column.ModifyColumnDefinitionFiller" />
+    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.column.alter.RenameColumnSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.column.RenameColumnDefinitionFiller" />
+    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.column.alter.DropColumnDefinitionSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.column.DropColumnDefinitionFiller" />
+    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.ddl.constraint.DropPrimaryKeySegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.ddl.constraint.DropPrimaryKeyFiller" />
     <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.dml.column.InsertColumnsSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.dml.InsertColumnsFiller" />
     <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.dml.SelectItemsSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.dml.SelectItemsFiller" />
     <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.dml.WhereSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.dml.WhereFiller" />
